refactor: move TypeScript transformer pipeline into src/Support/TypeScript
Relocates the seven transformer/provider/formatter classes from
workbench/app/Support (Workbench\App\Support) into the package src tree
(Lattice\Lattice\Support\TypeScript) so the pipeline is owned by the
package. Updates the workbench provider's imports accordingly; generated
output is unchanged.
