+++ src/Support/TypeScript/LatticeHttpMethodTransformer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Workbench\App\Support;
+namespace Lattice\Lattice\Support\TypeScript;
 
 use Lattice\Lattice\Core\Enums\HttpMethod;
 use Spatie\TypeScriptTransformer\Data\TransformationContext;
