+++ src/Support/TypeScript/LatticeValueObjectTransformer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Workbench\App\Support;
+namespace Lattice\Lattice\Support\TypeScript;
 
 use Spatie\TypeScriptTransformer\PhpNodes\PhpClassNode;
 use Spatie\TypeScriptTransformer\Transformers\ClassTransformer;
