+++ src/Support/TypeScript/OxfmtFormatter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Workbench\App\Support;
+namespace Lattice\Lattice\Support\TypeScript;
 
 use Spatie\TypeScriptTransformer\Formatters\Formatter;
 use Symfony\Component\Process\Exception\ProcessFailedException;
